refactor(webhook): merge minimal payload types into internal/webhook/payload
A full merge of internal/webhook/dto into internal/webhook/payload isn't
possible: internal/webhook/payload already imports internal/ee/service
(for its Services struct), and internal/ee/service needs the InternalXEvent
structs to publish webhook events -- moving those would create a cycle
(service -> payload -> service).

Splits the package instead:
- internal/webhook/dto keeps only what internal/ee/service actually needs:
  the ~13 InternalXEvent structs, plus WalletAlertInfo and RejectedEventData
  (two small publish-time helpers also constructed by the service layer).
- internal/webhook/payload gains everything else: every minimal entity type
  (Customer, Subscription, Invoice, ...), every envelope wrapper
  (CustomerWebhookPayload, InvoiceWebhookPayload, ...), and their
  constructors -- merged into the same file as the builder that uses them.

internal/api/v1/webhook_internal.go (the doc-only Swagger anchor file)
now imports internal/webhook/payload instead of internal/webhook/dto for
the types that moved.

docs/swagger/docs.go has stale "webhookDto.X" $ref strings until the next
`make swagger` run -- harmless (string literals in generated JSON, not
Go type references) but worth regenerating.
