+++ internal/webhook/dto/communication.go
@@ -1,20 +1,6 @@
 package webhookDto
 
-import (
-	"github.com/flexprice/flexprice/internal/api/dto"
-	"github.com/flexprice/flexprice/internal/types"
-)
-
 type InternalCommunicationEvent struct {
 	InvoiceID string `json:"invoice_id"`
 	TenantID  string `json:"tenant_id"`
-}
-
-type CommunicationWebhookPayload struct {
-	EventType types.WebhookEventName `json:"event_type"`
-	Invoice   *Invoice               `json:"invoice"`
-}
-
-func NewCommunicationWebhookPayload(invoice *dto.InvoiceResponse, eventType types.WebhookEventName) *CommunicationWebhookPayload {
-	return &CommunicationWebhookPayload{EventType: eventType, Invoice: NewInvoice(invoice, eventType)}
 }
