+++ internal/webhook/payload/alert.go
@@ -3,12 +3,112 @@ package payload
 import (
 	"context"
 	"encoding/json"
+	"time"
 
 	"github.com/flexprice/flexprice/internal/api/dto"
 	ierr "github.com/flexprice/flexprice/internal/errors"
 	"github.com/flexprice/flexprice/internal/types"
 	webhookDto "github.com/flexprice/flexprice/internal/webhook/dto"
+	"github.com/shopspring/decimal"
 )
+
+type AlertWebhookPayload struct {
+	EventType      types.WebhookEventName `json:"event_type"`
+	AlertType      types.AlertType        `json:"alert_type"`
+	AlertStatus    types.AlertState       `json:"alert_status"`
+	FeatureID      string                 `json:"feature_id,omitempty"`
+	WalletID       string                 `json:"wallet_id,omitempty"`
+	CustomerID     string                 `json:"customer_id,omitempty"`
+	CurrentBalance string                 `json:"current_balance,omitempty"`
+	CreditBalance  string                 `json:"credit_balance,omitempty"`
+}
+
+func NewAlertWebhookPayload(feature *dto.FeatureResponse, wallet *dto.WalletResponse, customer *dto.CustomerResponse, alertType types.AlertType, alertStatus types.AlertState, eventType types.WebhookEventName) *AlertWebhookPayload {
+	payload := &AlertWebhookPayload{
+		EventType:   eventType,
+		AlertType:   alertType,
+		AlertStatus: alertStatus,
+	}
+	if feature != nil && feature.Feature != nil {
+		payload.FeatureID = feature.ID
+	}
+	if wallet != nil && wallet.Wallet != nil {
+		payload.WalletID = wallet.ID
+		payload.CurrentBalance = wallet.Balance.String()
+		payload.CreditBalance = wallet.CreditBalance.String()
+	}
+	if customer != nil && customer.Customer != nil {
+		payload.CustomerID = customer.ID
+	}
+	return payload
+}
+
+type SpendAlertEvent struct {
+	Subscription           *Subscription    `json:"subscription"`
+	SubscriptionLineItemID string           `json:"subscription_line_item_id,omitempty"`
+	GroupID                string           `json:"group_id,omitempty"`
+	AlertType              types.AlertType  `json:"alert_type"`
+	AlertStatus            types.AlertState `json:"alert_status"`
+	CurrentSpend           string           `json:"current_spend"`
+	Threshold              *decimal.Decimal `json:"threshold,omitempty" swaggertype:"string"`
+	TriggeredAt            time.Time        `json:"triggered_at"`
+}
+
+func thresholdForAlertStatus(settings *types.AlertSettings, status types.AlertState) *decimal.Decimal {
+	if settings == nil {
+		return nil
+	}
+	var t *types.AlertThreshold
+	switch status {
+	case types.AlertStateInAlarm:
+		t = settings.Critical
+	case types.AlertStateWarning:
+		t = settings.Warning
+	case types.AlertStateInfo:
+		t = settings.Info
+	}
+	if t == nil {
+		return nil
+	}
+	return &t.Threshold
+}
+
+func NewSpendAlertEvent(sub *dto.SubscriptionResponse, lineItemID, groupID string, alertType types.AlertType, alertStatus types.AlertState, currentSpend string, alertSettings *types.AlertSettings, triggeredAt time.Time) *SpendAlertEvent {
+	return &SpendAlertEvent{
+		Subscription:           NewSubscription(sub),
+		SubscriptionLineItemID: lineItemID,
+		GroupID:                groupID,
+		AlertType:              alertType,
+		AlertStatus:            alertStatus,
+		CurrentSpend:           currentSpend,
+		Threshold:              thresholdForAlertStatus(alertSettings, alertStatus),
+		TriggeredAt:            triggeredAt,
+	}
+}
+
+type EntitlementGrantAlertEvent struct {
+	SubscriptionID     string           `json:"subscription_id"`
+	CustomerID         string           `json:"customer_id"`
+	EntitlementID      string           `json:"entitlement_id"`
+	EntitlementGrantID string           `json:"entitlement_grant_id"`
+	AlertType          types.AlertType  `json:"alert_type"`
+	AlertStatus        types.AlertState `json:"alert_status"`
+	UsageRatio         string           `json:"usage_ratio"`
+	TriggeredAt        time.Time        `json:"triggered_at"`
+}
+
+func NewEntitlementGrantAlertEvent(subscriptionID, customerID, entitlementID, grantID string, alertType types.AlertType, alertStatus types.AlertState, usageRatio string, triggeredAt time.Time) *EntitlementGrantAlertEvent {
+	return &EntitlementGrantAlertEvent{
+		SubscriptionID:     subscriptionID,
+		CustomerID:         customerID,
+		EntitlementID:      entitlementID,
+		EntitlementGrantID: grantID,
+		AlertType:          alertType,
+		AlertStatus:        alertStatus,
+		UsageRatio:         usageRatio,
+		TriggeredAt:        triggeredAt,
+	}
+}
 
 type AlertPayloadBuilder struct {
 	services *Services
@@ -65,7 +165,7 @@ func (b *AlertPayloadBuilder) BuildPayload(ctx context.Context, eventType types.
 		}
 
 		// Build the complete alert webhook payload with both entities and customer
-		payload := webhookDto.NewAlertWebhookPayload(
+		payload := NewAlertWebhookPayload(
 			feature,
 			wallet,
 			customer,
@@ -81,6 +181,10 @@ func (b *AlertPayloadBuilder) BuildPayload(ctx context.Context, eventType types.
 	return nil, nil
 }
 
+// buildEntitlementGrantAlertPayload resolves a grant-exhaustion alert into its webhook payload.
+// Only the grant itself is fetched -- subscription/customer/entitlement IDs are already known
+// from internalEvent.ParentEntityID and the grant's own CustomerID/EntitlementConfigID fields,
+// so under the ID-only payload policy no further fetches are needed.
 func (b *AlertPayloadBuilder) buildEntitlementGrantAlertPayload(ctx context.Context, internalEvent webhookDto.InternalAlertEvent) (json.RawMessage, error) {
 	if internalEvent.ParentEntityID == "" {
 		return nil, ierr.NewError("entitlement grant alert missing subscription id").
@@ -93,7 +197,7 @@ func (b *AlertPayloadBuilder) buildEntitlementGrantAlertPayload(ctx context.Cont
 		return nil, err
 	}
 
-	payload := webhookDto.NewEntitlementGrantAlertEvent(
+	payload := NewEntitlementGrantAlertEvent(
 		internalEvent.ParentEntityID,
 		grant.CustomerID,
 		grant.EntitlementConfigID,
@@ -106,6 +210,9 @@ func (b *AlertPayloadBuilder) buildEntitlementGrantAlertPayload(ctx context.Cont
 	return json.Marshal(payload)
 }
 
+// buildSpendAlertPayload resolves an InternalAlertEvent carrying a subscription/line-item/group
+// spend alert into its final webhook payload. Only the subscription is fetched -- the line item
+// or group is represented by its already-known ID (internalEvent.EntityID), not fetched in full.
 func (b *AlertPayloadBuilder) buildSpendAlertPayload(ctx context.Context, internalEvent webhookDto.InternalAlertEvent) (json.RawMessage, error) {
 	// A line-item or group alert's entity_id is the line item/group itself; the subscription it
 	// rolls up to is parent_entity_id. A subscription-level alert has no parent, so entity_id is
@@ -128,7 +235,7 @@ func (b *AlertPayloadBuilder) buildSpendAlertPayload(ctx context.Context, intern
 		groupID = internalEvent.EntityID
 	}
 
-	payload := webhookDto.NewSpendAlertEvent(
+	payload := NewSpendAlertEvent(
 		sub, lineItemID, groupID,
 		internalEvent.AlertType, internalEvent.AlertStatus,
 		internalEvent.AlertInfo.ValueAtTime.String(), internalEvent.AlertInfo.AlertSettings, internalEvent.AlertInfo.Timestamp,
