+++ internal/webhook/payload/subscription_phase.go
@@ -3,11 +3,41 @@ package payload
 import (
 	"context"
 	"encoding/json"
+	"time"
 
+	"github.com/flexprice/flexprice/internal/api/dto"
 	ierr "github.com/flexprice/flexprice/internal/errors"
 	"github.com/flexprice/flexprice/internal/types"
 	webhookDto "github.com/flexprice/flexprice/internal/webhook/dto"
 )
+
+type SubscriptionPhase struct {
+	ID             string     `json:"id"`
+	SubscriptionID string     `json:"subscription_id"`
+	StartDate      time.Time  `json:"start_date"`
+	EndDate        *time.Time `json:"end_date,omitempty"`
+}
+
+func NewSubscriptionPhase(resp *dto.SubscriptionPhaseResponse) *SubscriptionPhase {
+	if resp == nil || resp.SubscriptionPhase == nil {
+		return nil
+	}
+	return &SubscriptionPhase{
+		ID:             resp.ID,
+		SubscriptionID: resp.SubscriptionID,
+		StartDate:      resp.StartDate,
+		EndDate:        resp.EndDate,
+	}
+}
+
+type SubscriptionPhaseWebhookPayload struct {
+	EventType types.WebhookEventName `json:"event_type"`
+	Phase     *SubscriptionPhase     `json:"phase"`
+}
+
+func NewSubscriptionPhaseWebhookPayload(phase *SubscriptionPhase, eventType types.WebhookEventName) *SubscriptionPhaseWebhookPayload {
+	return &SubscriptionPhaseWebhookPayload{EventType: eventType, Phase: phase}
+}
 
 type SubscriptionPhasePayloadBuilder struct {
 	services *Services
@@ -37,7 +67,7 @@ func (b *SubscriptionPhasePayloadBuilder) BuildPayload(ctx context.Context, even
 		return nil, err
 	}
 
-	payload := webhookDto.NewSubscriptionPhaseWebhookPayload(webhookDto.NewSubscriptionPhase(phase), eventType)
+	payload := NewSubscriptionPhaseWebhookPayload(NewSubscriptionPhase(phase), eventType)
 
 	return json.Marshal(payload)
 }
